Ensure audit flexibility
Due to the lack of existing S/MIME criteria, and the wide variety of schemes and requirements, this ensures a flexible participation model for CAs.

In addition, it removes the use of the term "Qualified Auditor", which, although part of our Bylaws, refers back to the SCWG's Baseline Requirements. As this is a new CWG, we're not quite ready to take in a dependency on a different CWG, especially within the CWG's charter.