Additional cleanups and markdownlint compliance
This commit updates the BR.md doc with the following changes:
* Removes all instances of two adjacent spaces, except when used
  for indentation
* Removes all instances of two adjacent blank lines
* Fixes all markdownlints, including spaces around lists, spaces
  around headings, and tags on fenced blocks.

It also fixes a few other pieces of miscellany, such as paragraphs
which were not sufficiently indented to semantically fall inside the
list in which they are intended to fall, and fencing code blocks which
should have been fenced.

It does not address two markdownlints: There are still multiple H1
headings, and there are a few places where markdown believes that
unordered list bullets should be unindented (but doing so would
change the semantics of the document).